Implement `BoxUnboxVmValue` for `spinoso_string::String`
Implement a custom mruby boxing and unboxing routine for
`spinoso_string::String`. This trait implementation is similar to the
one for `Array`: decompose the `String`'s backing `Vec` into "raw
parts". Then pack the resulting `Vec` triple into the `*mut sys::RString`.
The `String`'s encoding is stored in the `*mut sys::RBasic` flags. Care
is taken to ensure the encoding flag bitpos does not clobber any native
mruby flags such as the "is frozen?" bit.

This commit adds two functions to `mruby-sys`'s `ext.c` for allocating
an `RString *` from "raw parts" and repacking "raw parts" into an
`mrb_value`. These functions mirror the ones introduced for this purpose
in `Array`'s `BoxUnboxVmValue` implementation. The unions are cleaner to
interact with in C and the `RSTRING` macro makes the code easier to
read.